+++ chat_misc_integration_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 	"time"
 
-	. "github.com/GetStream/getstream-go/v3"
+	. "github.com/GetStream/getstream-go/v4"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )