+++ chat_user_integration_test.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"testing"
 
-	. "github.com/GetStream/getstream-go/v3"
+	. "github.com/GetStream/getstream-go/v4"
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"