+++ chat_webhook_test.go
@@ -9,7 +9,7 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	. "github.com/GetStream/getstream-go/v3"
+	. "github.com/GetStream/getstream-go/v4"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )